Cleaned up and removed legacy traversal junk
