+++ core/src/main/scala/lightdb/traversal/DocumentTraversal.scala
@@ -11,11 +11,3 @@ case class DocumentTraversal[Doc <: Document[Doc]](ids: rapid.Stream[Id[Doc]]) {
     EdgeTraversal(edges)
   }
 }
-
-case class EdgeTraversal[E <: EdgeDocument[E, F, T], F <: Document[F], T <: Document[T]](edges: rapid.Stream[E]) {
-  def follow[E2 <: EdgeDocument[E2, T, T2], T2 <: Document[T2]](tx: PrefixScanningTransaction[E2, _]): EdgeTraversal[E2, T, T2] = {
-    val nextEdges = edges.flatMap(edge => tx.traversal.edgesFor[E2, T, T2](edge._to))
-    EdgeTraversal(nextEdges)
-  }
-  def documents(tx: Transaction[T, _]): rapid.Stream[T] = edges.evalMap(edge => tx(edge._to))
-}
