+++ core/src/main/scala/lightdb/traversal/TraversalBuilder.scala
@@ -177,38 +177,3 @@ object TraversalBuilder {
     new EdgeTraversalBuilder[E, From, To](fromId, transaction)
   }
 }
-
-/**
- * Specialized builder for edge traversal
- */
-class EdgeTraversalBuilder[E <: EdgeDocument[E, From, To], From <: Document[From], To <: Document[To]](
-                                                                                                        startId: Id[From],
-                                                                                                        transaction: PrefixScanningTransaction[E, _]
-                                                                                                      ) {
-  /**
-   * Follow edges using the provided step
-   */
-  def via[M <: DocumentModel[E]](
-                                  step: GraphStep[E, M, From, To]
-                                ): Task[Seq[Id[To]]] = {
-    step.neighbors(startId, transaction.asInstanceOf[PrefixScanningTransaction[E, M]])
-      .map(_.toSeq)
-  }
-
-  /**
-   * Get a stream of results
-   */
-  def stream: Stream[Id[To]] = {
-    // Use prefixStream directly for efficiency
-    transaction.prefixStream(startId.value)
-      .filter(_._from == startId)
-      .map(_._to)
-  }
-
-  /**
-   * Get all results as a list
-   */
-  def toList: Task[List[Id[To]]] = {
-    stream.toList
-  }
-}
